fix bug where global vars would be preferred over local vars
before this fix the compiler would choose the global variable with the same name, if it existed, over the local variable which is not how it should work.
This fix changes the this. No the compiler first checks, if a local variable with the name exists and then if no local variable with the name was found it checks if a global variable with the name exists.