simplified func-param gen + fix division issue
when dividing, RAX contains the quotient and RDX contains the remainder. RDX also contains the high bits of the dividend (RDX:RAX).
The problem was, that the compiler did not know that RDX could be non zero.
This lead to the problem, where RDX should be zero, but is set to the remainder of an other division, which results in the wrong result or an FloatingPoint exception -> crash of the program
Implemented Solution: clear RDX before every division with XOR (xor rdx, rdx)
Better solution: the compiler should know if RDX could be non zero and then clear RDX before the division